requirements.txt
aiogram==3.13.1
aiohttp>=3.10,<3.11
fastapi==0.111.0
uvicorn==0.30.3
httpx==0.27.0
python-dotenv==1.0.1